==> Code/03_Data_Cleaning.ipynb <==
import pandas as pd
import numpy as np
from textblob import TextBlob
import regex as re
from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer 
from sklearn.decomposition import PCA
from sklearn.preprocessing import StandardScaler
import pickle

df = pd.read_csv('../Data/train.csv')

df['text'] = df['text'].astype(str)

# Removing all punctuation 
df['text'] = df['text'].map(lambda x: re.sub('[^\w\s]', ' ', x)) 

#Only keeping letters 
df['text'] = df['text'].map(lambda x: re.sub("[^a-zA-Z]", " ", x)) 

___

## Textblob Subjectivity and Polarity

def sentiment(x):
    try:
        return TextBlob(x).sentiment.subjectivity
    except:
        return 10
#10 is placeholder in case of an error

def polarity(x):
    try:
        return TextBlob(x).sentiment.polarity
    except:
        return 10
#10 is placeholder in case of an error

df['subjectivity'] = df['text'].apply(lambda x: sentiment(x))

df['polarity'] = df['text'].apply(lambda x: polarity(x))

df.index = range(df.shape[0])
df.head()

___

## CountVectorization

#using countervectorizer to tokenize our texts into collection of words 
cvec = CountVectorizer(stop_words = 'english', min_df = 10, max_df = 1.0,
                       ngram_range = (1,2))
words_matrix = cvec.fit_transform(df['text'])

# Saving the cvec as a pickle
filename = '../Assets/cvec.pkl'
pickle.dump(cvec, open(filename, 'wb'))

#creating a datframe from all tokenized words  
words_df = pd.DataFrame(words_matrix.toarray(), columns=cvec.get_feature_names())

print(words_df.shape)
print(df[['subjectivity', 'polarity']].shape)

words_df['subjectivity'] = df['subjectivity']
words_df['polarity'] = df['polarity']
words_df['target'] = df['target']

words_df[['subjectivity', 'polarity','target']].corr()

words_df.head()

## TFIDFVectorizer

#using countervectorizer to tokenize our texts into collection of words 
tfidf = TfidfVectorizer(stop_words = 'english', min_df = 15, max_df = 1.0,
                       ngram_range = (1,2))
words_matrix1 = tfidf.fit_transform(df['text'])

# Saving tfidf as a pickle
filename = '../Assets/tfidf.pkl'
pickle.dump(tfidf, open(filename, 'wb'))

#creating a datframe from all tokenized words  
words_df1 = pd.DataFrame(words_matrix1.toarray(), columns=tfidf.get_feature_names())

words_df1['subjectivity'] = df['subjectivity']
words_df1['polarity'] = df['polarity']
words_df1['target'] = df['target']

words_df1.shape

## Saving to csv

words_df.to_csv('../Data/words_df.csv', index = False)
words_df1.to_csv('../Data/words_df1.csv', index = False)
