==> Code/04_Modeling.ipynb <==
import pandas as pd
import numpy as np
import pickle
from sklearn.model_selection import train_test_split, GridSearchCV, cross_val_score, cross_val_predict
from sklearn.preprocessing import StandardScaler
from sklearn.decomposition import PCA
from sklearn.linear_model import LogisticRegression
from sklearn.ensemble import RandomForestClassifier, AdaBoostClassifier
from sklearn.naive_bayes import MultinomialNB, GaussianNB
from sklearn.metrics import precision_score, roc_auc_score, recall_score
from sklearn.tree import DecisionTreeClassifier

### Reading in cleaned datasets

words_df = pd.read_csv('../Data/words_df.csv')
words_df1 = pd.read_csv('../Data/words_df1.csv')

## PCA

#countvetorized
X = words_df.drop(columns = 'target')
y = words_df['target']

#tfidif
X1 = words_df1.drop(columns = 'target')
y1 = words_df1['target']

words_df.head()

#no NLP
X2 = words_df.drop(columns = ['target', 'subjectivity', 'polarity'])
y2 = words_df['target']

y.value_counts()

X_train, X_test, y_train, y_test = train_test_split(X, y, random_state=42, stratify = y)

X_train1, X_test1, y_train1, y_test1 = train_test_split(X1, y1, random_state=42, stratify = y)

ss = StandardScaler()
X_train = ss.fit_transform(X_train)
X_test = ss.transform(X_test)

from sklearn.decomposition import PCA

pca = PCA()
pca.fit(X_train)
Z_train = pca.transform(X_train)
Z_test = pca.transform(X_test)

var_exp = pca.explained_variance_ratio_
cum_var_exp = np.cumsum(var_exp)
cum_var_exp

# Modeling

logreg =LogisticRegression(solver = 'liblinear')

params_logreg = {'penalty': ['l1', 'l2']}

### Logisitic Regression with Scaling

gs_logreg = GridSearchCV(logreg, params_logreg, cv = 5)
gs_logreg.fit(X_train, y_train)

# Model Evaluation
# print out the best score
print(gs_logreg.best_score_)
#print out the set of hyperparameters that achieved the best score.
logreg_model = gs_logreg.best_estimator_
print('accuracy score - Existing data', logreg_model.score(X_train, y_train))
print('accuracy score - New data',  logreg_model.score(X_test, y_test)) # Accuracy score 
print(roc_auc_score(y_test, logreg_model.predict_proba(X_test)[:,1]))
print(precision_score(y_test, logreg_model.predict(X_test)))

file_name = '../Assets/logreg.sav'
pickle.dump(logreg_model, open(file_name, 'wb'))

### Logisitic Regressio with Tfidf + Scaling

gs_logreg1 = GridSearchCV(logreg, params_logreg, cv = 5)
gs_logreg1.fit(X_train1, y_train1)

# Model Evaluation
# print out the best score
print(gs_logreg1.best_score_)
#print out the set of hyperparameters that achieved the best score.
logreg_model1 = gs_logreg1.best_estimator_
print('accuracy score - Existing data', logreg_model1.score(X_train1, y_train1))
print('accuracy score - New data',  logreg_model1.score(X_test1, y_test1)) # Accuracy score 
print(roc_auc_score(y_test, logreg_model1.predict_proba(X_test1)[:,1]))
print(precision_score(y_test, logreg_model1.predict(X_test1)))

file_name = '../Assets/logreg_tf_ss.sav'
pickle.dump(logreg_model1, open(file_name, 'wb'))

### Logisitic Regression with Scaling + PCA

#gs_logreg1 = GridSearchCV(logreg, params_logreg, cv = 5)
#gs_logreg1.fit(Z_train, y_train)

# Model Evaluation
# print out the best score
#print(gs_logreg1.best_score_)
#print out the set of hyperparameters that achieved the best score.
#logreg_model1 = gs_logreg1.best_estimator_
#print('accuracy score - Existing data', logreg_model1.score(Z_train, y_train))
#print('accuracy score - New data',  logreg_model1.score(Z_test, y_test)) # Accuracy score 

___

### Randomforest with Scaling

rf = RandomForestClassifier()

rf_params = {
    'n_estimators': [100, 150, 200],
    'max_depth': [None, 5, 10], 
    'min_samples_split' : [2,3, 5]
}
gs_rf = GridSearchCV(rf, param_grid=rf_params, cv=5)
gs_rf.fit(X_train, y_train)

print(gs_rf.best_score_)
#print out the set of hyperparameters that achieved the best score.
rf_model = gs_rf.best_estimator_
print('accuracy score - Existing data', rf_model.score(X_train, y_train))
print('accuracy score - New data', rf_model.score(X_test, y_test)) # Accuracy 

roc_auc_score(y_test, rf_model.predict_proba(X_test)[:,1])

precision_score(y_test, rf_model.predict(X_test))

file_name = '../Assets/rf_ss.sav'
pickle.dump(rf_model, open(file_name, 'wb'))

### Randomforest with Scaling + Tfidf

gs_rf1 = GridSearchCV(rf, param_grid=rf_params, cv=5)
gs_rf1.fit(X_train1, y_train1)

print(gs_rf1.best_score_)
#print out the set of hyperparameters that achieved the best score.
rf_model1 = gs_rf1.best_estimator_
print('accuracy score - Existing data', rf_model1.score(X_train1, y_train1))
print('accuracy score - New data', rf_model1.score(X_test1, y_test1)) # Accuracy 

roc_auc_score(y_test1, rf_model1.predict_proba(X_test1)[:,1])

precision_score(y_test1, rf_model1.predict(X_test1))

file_name = '../Assets/rf_ss_tf.sav'
pickle.dump(rf_model1, open(file_name, 'wb'))

### Ada Boost with tfidf

ada = AdaBoostClassifier(base_estimator=DecisionTreeClassifier())

ada_params = {
    'n_estimators': [50,100],
    'base_estimator__max_depth': [1,2],
    'learning_rate': [.9, 1.]
}

gs_ada = GridSearchCV(ada, param_grid=ada_params, cv=5)
gs_ada.fit(X_train1, y_train1)

print(gs_ada.best_score_)
#print out the set of hyperparameters that achieved the best score.
ada_model = gs_ada.best_estimator_
print('accuracy score - Existing data', ada_model.score(X_train1, y_train1))
print('accuracy score - New data', ada_model.score(X_test1, y_test1)) # Accuracy 

roc_auc_score(y_test1, ada_model.predict_proba(X_test1)[:,1])

precision_score(y_test1, ada_model.predict(X_test1))

file_name = '../Assets/ada_tf.sav'
pickle.dump(ada_model, open(file_name, 'wb'))

## Multinomial Naive Beyes

X_train2, X_test2, y_train2, y_test2 = train_test_split(X2, y2, random_state=42, stratify = y)

file_name = '../Assets/words_df_columns.pkl'
pickle.dump(X_train2.columns, open(file_name, 'wb'))

mnb = MultinomialNB()

mnb_model = mnb.fit(X_train2, y_train2)

print(mnb_model.score(X_train2, y_train2))
print(mnb_model.score(X_test2, y_test2))

roc_auc_score(y_test2, mnb_model.predict_proba(X_test2)[:,1])

precision_score(y_test2, mnb_model.predict(X_test2))

file_name = '../Assets/multi_bayes.sav'
pickle.dump(mnb_model, open(file_name, 'wb'))

## Gaussian NB

gnb = GaussianNB()
gnb_model = gnb.fit(X_train2, y_train2)

print(gnb_model.score(X_train2, y_train2))
print(gnb_model.score(X_test2, y_test2))

roc_auc_score(y_test2, gnb_model.predict_proba(X_test2)[:,1])

precision_score(y_test2, gnb_model.predict(X_test2))

file_name = '../Assets/gaussian.sav'
pickle.dump(gnb_model, open(file_name, 'wb'))

## Metrics

import pickle

gnb_model = pickle.load(open('../Assets/gaussian.sav', 'rb'))
mnb_model = pickle.load(open('../Assets/multi_bayes.sav', 'rb'))
ada_model = pickle.load(open('../Assets/ada_tf.sav', 'rb'))
logreg_model = pickle.load(open('../Assets/logreg.sav', 'rb'))
logreg_model1 = pickle.load(open('../Assets/logreg_tf_ss.sav', 'rb'))
rf_model = pickle.load(open('../Assets/rf_ss.sav', 'rb'))
rf_model1 = pickle.load(open('../Assets/rf_ss_tf.sav', 'rb'))
