==> Code/05_Twitter_Application.ipynb <==
import pandas as pd
import numpy as np
import pickle
from textblob import TextBlob
from sklearn.model_selection import train_test_split, GridSearchCV, cross_val_score, cross_val_predict
from sklearn.preprocessing import StandardScaler
from sklearn.decomposition import PCA
from sklearn.linear_model import LogisticRegression 
from sklearn.ensemble import RandomForestClassifier, AdaBoostClassifier
from sklearn.naive_bayes import MultinomialNB, GaussianNB
from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer 

## Twitter Data

twitter = pd.read_csv('../Data/twitter.csv')

twitter['text'].dropna(inplace = True)

### Count Vectorizing Twitter

cvec = pickle.load(open('../Assets/cvec.pkl', 'rb'))

twitter_cvec = cvec.transform(twitter['text'])

twitter_cvec_df = pd.DataFrame(twitter_cvec.toarray(), columns=cvec.get_feature_names())

# There are 17 too many columns here... Should be 1690
twitter_cvec_df.head()

### 

pickle_columns = pickle.load(open('../Assets/words_df_columns.pkl', 'rb'))

twitter_cvec_df = twitter_cvec_df[pickle_columns]

### TFIDF Vectorizing Twitter

tfidf = pickle.load(open('../Assets/tfidf.pkl', 'rb'))

twitter_tfidf = tfidf.transform(twitter['text'])

twitter_tfidf_df = pd.DataFrame(twitter_tfidf.toarray(), columns = tfidf.get_feature_names())

### Modeling

mnb = pickle.load(open('../Assets/multi_bayes.sav', 'rb'))

twitter_cvec_df.shape

cvec_preds = mnb.predict(twitter_cvec_df)

twitter_preds = twitter.dropna(subset = ['text']).copy()
twitter_preds['prediction'] = cvec_preds
twitter_preds = twitter_preds[['text','prediction']]
twitter_preds['prediction'] = twitter_preds['prediction'].map({0:'Not Reliable', 1:'Reliable'})

twitter_preds
